Consolidate webhooks commands into single file
